chore: move last function in helpers file to the modifier
